+++ utils/configuration.py
@@ -25,20 +25,20 @@ if not Path(CONFIG_PATHS['user']).is_file():
         f.write('')
 
 with open(CONFIG_PATHS['default'], 'r', encoding='utf-8') as f:
-    CONFIG_DEFAULT = benedict(yaml.safe_load(f))
+    CONFIG_DEFAULT_DICT = benedict(yaml.safe_load(f))
 
 with open(CONFIG_PATHS['user'], 'r', encoding='utf-8') as f:
     CONFIG_YAML = yaml.safe_load(f)
-    CONFIG = benedict(CONFIG_YAML) if CONFIG_YAML else benedict({})
+    CONFIG_DICT = benedict(CONFIG_YAML) if CONFIG_YAML else benedict({})
 
 def get(key: str) -> Any:
     """Looks for the given key in the user config, returns the default value if none is set."""
-    return CONFIG.get(key, CONFIG_DEFAULT.get(key))
+    return CONFIG_DICT.get(key, CONFIG_DEFAULT_DICT.get(key))
 
 def get_default(key: str) -> Any:
     """Returns the default value for a given key."""
-    return CONFIG_DEFAULT.get(key)
+    return CONFIG_DEFAULT_DICT.get(key)
 
 def get_full(config_type: Literal['user', 'default']) -> benedict:
     """Returns the full config benedict object."""
-    return CONFIG if config_type == 'user' else CONFIG_DEFAULT
+    return CONFIG_DICT if config_type == 'user' else CONFIG_DEFAULT_DICT