+++ docs/changelog.md
@@ -4,12 +4,23 @@
 
 > *2024.mm.dd / dev.34*
 
+
+
 Developer
 - Docstrings have been added to most functions, classes, and modules
 - Any instances of `import regex as re` have been replaced with `import re`
 - `customlog.py` has been removed entirely, logging is now handled using the [`colorlog`](https://pypi.org/project/colorlog/) library
     - Therefore, logs are no longer marked as `verbose` with a keyword argument, and instead use standard logging levels. Most logs that used `verbose` have been moved to `DEBUG`-level logs, but some have been deemed `INFO`-level instead
-- `utils` directory added to contain helper modules
+- Changes in `bot.py`:
+    - `debugctx` variable and `-dctx` command removed, test commands will grab a `Context` object automatically
+    - Many `global` statements have been removed, various variables moved into separate cogs or classes where relevant
+    - Cogs have been moved into separate files, located in `cogs/`
+        - `General` cog moved into `cog_general.py`
+        - `Voice` cog moved into `cog_voice.py`, along with most functions and classes related to voice connection and audio playback; other changes have been made within this file, such as...
+            - `MediaQueue` no longer keeps track of multiple queues per Discord server and instead represents just a single queue (part of [#52](https://github.com/svioletg/viMusBot/issues/52))
+                - It also now contains things like `now_playing`, `last_played`, `is_looping` (formerly `loop_this`), etc.
+    - `embedq()` no longer uses `*args`, now has proper keyword arguments — `title` (`str`; main, largest text), `subtext` (`str`; shown below `title` in smaller font), and `color` (`int`)
+- `utils/` directory added to contain helper modules
     - `miscutil.py` created in this directory to house general-purpose utility methods that should be shared between modules
     - `configuration.py` created in this directory to reduce the amount of duplicated code regarding configuration across this project
         - This module has a `get()` function that automatically retrieves the default value if none is set in the custom configuration, this removes the need for every single file to have the key typed out twice, e.g. `config.get('allow-spotify-playlists', config_default['allow-spotify-playlists'])`, and can now just be `config.get('allow-spotify-playlists')`